requirements.txt
boto3==1.5.28
botocore==1.8.43
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.1.0
django-dotenv==1.4.2
django-extensions==1.9.9
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.7
djangorestframework-camel-case==0.2.0
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
typing==3.6.4
